refactoring and fixing components and functionality according home task 3

+++ src/services.js
@@ -18,11 +18,27 @@ export async function register(name, email, password) {
 }
 
 export async function createCourse(courseData) {
+	console.log('course data: ', courseData);
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
+
 	const response = await fetch('http://localhost:4000/courses/add', {
 		method: 'POST',
 		body: JSON.stringify(courseData),
-		headers: { 'Content-Type': 'application/json' },
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
 	});
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
 	return await response.json();
 }
 
@@ -36,12 +52,30 @@ export async function getCourses() {
 }
 
 export async function getCourse(id) {
-	const response = await fetch('http://localhost:4000/courses/{id}', {
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	console.log('course for loading: ', id);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
+	const URL = `http://localhost:4000/courses/${id}`;
+	const response = await fetch(URL, {
 		method: 'GET',
-		headers: { 'Content-Type': 'application/json' },
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
 	});
-	const data = await response.json();
-	return data;
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
+	if (response.ok) {
+		console.log('Course item with id' + { id } + 'successfully loaded!');
+	}
+	return response.json();
 }
 
 export async function editCourse(id, courseData) {
@@ -54,12 +88,30 @@ export async function editCourse(id, courseData) {
 }
 
 export async function deleteCourse(id) {
-	const response = await fetch('http://localhost:4000/courses/{id}', {
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	console.log('course for deleting: ', id);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
+	const URL = `http://localhost:4000/courses/${id}`;
+	const response = await fetch(URL, {
 		method: 'DELETE',
-		headers: { 'Content-Type': 'application/json' },
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
 	});
-	const data = await response.json();
-	return data;
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
+	if (response.ok) {
+		console.log('Course item with id' + { id } + 'successfully deleted!');
+	}
+	return await response.json();
 }
 
 export async function getAuthors() {
@@ -72,19 +124,74 @@ export async function getAuthors() {
 }
 
 export async function createAuthor(authorData) {
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
 	const response = await fetch('http://localhost:4000/authors/add', {
 		method: 'POST',
 		body: JSON.stringify(authorData),
-		headers: { 'Content-Type': 'application/json' },
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
 	});
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
 	return await response.json();
 }
 
 export async function deleteAuthor(id) {
-	const response = await fetch('http://localhost:4000/authors/{id}', {
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	console.log('author for deleting: ', id);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
+	const URL = `http://localhost:4000/authors/${id}`;
+	const response = await fetch(URL, {
 		method: 'DELETE',
-		headers: { 'Content-Type': 'application/json' },
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
 	});
-	const data = await response.json();
-	return data;
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
+	if (response.ok) {
+		console.log('item with id' + { id } + 'successfully deleted!');
+	}
+	return await response.json();
+}
+
+export async function editAuthor(id, authorData) {
+	const accessToken = localStorage.getItem('result');
+	const token = accessToken.split(' ')[1].slice(0, -2);
+	if (!accessToken) {
+		throw new Error('Access token not found');
+	}
+	const response = await fetch('http://localhost:4000/authors/' + { id }, {
+		method: 'PUT',
+		body: JSON.stringify(authorData),
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`,
+		},
+	});
+	if (response.status === 401) {
+		throw new Error('Unauthorized');
+	}
+	if (!response.ok) {
+		throw new Error(`Network response was not ok (${response.status})`);
+	}
+	return await response.json();
 }
